pyproject.toml
[project]
name = "service-url-shortener"
version = "0.0.1"
description = ""
authors = [{ name = "Josh Hsieh", email = "jtr860830@gmail.com" }]
dependencies = [
    "fastapi>=0.83.0",
    "uvicorn[standard]>=0.18.3",
    "nanoid>=2.0.0",
    "sqlalchemy>=1.4.41",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "alembic>=1.8.1",
    "pg8000>=1.29.1",
    "pydantic>=1.10.2",
]
requires-python = ">=3.10"
license = { text = "MIT" }
[project.optional-dependencies]

[tool.pdm]
[tool.pdm.dev-dependencies]
test = ["pytest>=7.1.3"]
lint = ["flake8>=5.0.4"]
format = ["black>=22.8.0"]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
